refactor: Add PureTransaction and PureBlock (no tx hash)
Those new types are currently unused. The underlying class has no
transaction hash compiled in. So it is safe to use wherever it compiles.

The existing CMutableTransaction retains the behavior to calculate the
tx hash on demand. Use of CMutableTransaction makes compile-time
complexity analysis impossible.

The existing CTransaction retains the behavior to cache the hash on
initialization.